+++ frontend/src/App.jsx
@@ -1,21 +1,21 @@
 import React from "react";
 
 import { BrowserRouter, Routes, Route, useNavigate } from "react-router-dom";
-import Account from "./Account/accountPage";
+import Account from "./components/Account/accountPage";
 
-import Navbar from "./Navbar/Navbar_Landing";
-import { Main } from "./Main/mainUserPage";
-import Login from "./Login/login";
-import SignUp from "./SignUp/signUp";
-import ConfirmRegister from "./SignUp/ConfirmRegister";
+import Navbar from "./components/Navbar/Navbar_Landing";
+import { Main } from "./components/Main/mainUserPage";
+import Login from "./components/Login/login";
+import SignUp from "./components/SignUp/signUp";
+import ConfirmRegister from "./components/SignUp/ConfirmRegister";
 // import PrivateRoute from "./PrivateRoute";
 import { withAuthenticator } from "@aws-amplify/ui-react";
 import "@aws-amplify/ui-react/styles.css";
 import awsconfig from "./aws-exports";
 import { Amplify } from "aws-amplify";
-import Navbar_Landing from "./Navbar/Navbar_Landing";
-import Navbar_Home from "./Navbar/Navbar_Home";
-import Landing from "./Landing/Landing";
+import Navbar_Landing from "./components/Navbar/Navbar_Landing";
+import Navbar_Home from "./components/Navbar/Navbar_Home";
+import Landing from "./components/Landing/Landing";
 
 Amplify.configure(awsconfig);
