+++ frontend/src/components/Account/accountPage.jsx
@@ -4,7 +4,7 @@ import { Auth } from "aws-amplify";
 import { Storage } from "@aws-amplify/storage"
 import AWS from "aws-sdk";
 import {Buffer} from "buffer";
-import Submission from "../../Main/Submission/submission";
+import Submission from "../Main/Submission/submission";
 import { useNavigate } from "react-router-dom";
 import { Link } from "react-router-dom";
 import { Modal } from "react-bootstrap";
